Move the instances from Product, Choice and Codiagonal to Data.Lens.Common and Data.Lens.Patial.Common respectively.
