+++ src/Data/Lens/Common.hs
@@ -30,6 +30,9 @@ module Data.Lens.Common
 import Control.Applicative
 import Control.Comonad.Trans.Store
 import Control.Category
+import Control.Category.Choice
+import Control.Category.Product
+import Control.Category.Codiagonal
 import Data.Functor.Identity
 import Data.Functor.Apply
 import Data.Semigroupoid
@@ -151,3 +154,19 @@ intSetLens :: Int -> Lens IntSet Bool
 intSetLens k = Lens $ \m -> store (\mv ->
     if mv then IntSet.insert k m else IntSet.delete k m
   ) (IntSet.member k m)
+
+instance Choice Lens where
+  Lens f ||| Lens g =
+    Lens $ either
+      (\a -> let x = f a in store (Left . flip peek x) (pos x))
+      (\b -> let y = g b in store (Right . flip peek y) (pos y))
+
+instance Product Lens where
+  Lens f *** Lens g =
+    Lens $ \(a, c) ->
+      let x = f a
+          y = g c
+      in store (\(b, d) -> (peek b x, peek d y)) (pos x, pos y)
+
+instance Codiagonal Lens where
+  codiagonal = id ||| id
